move to a separate router for the app
