+++ template/src/{{ project_slug }}/frontend/default_routes/auth.py
