+++ template/src/{{ project_slug }}/frontend/default_routes/debug.py
